[UPDATE] convert the exeggutor code to a callable package
