+++ executor/stats.go
@@ -1,0 +1,8 @@
+package executor
+
+type Stats struct {
+	CpuTime     float64
+	ElapsedTime float64
+	Memory      int
+	Mark        string
+}
