+++ go.mod
@@ -3,3 +3,5 @@ module github.com/kumanik5661/exeggutor
 go 1.19
 
 require golang.org/x/sys v0.4.0
+
+require github.com/rabbitmq/amqp091-go v1.8.1 // indirect
